+++ Core/Rendering/mitkPointSetMapper2D.h
@@ -74,7 +74,7 @@ protected:
   bool m_Polygon;
   bool m_ShowPoints;
   bool m_ShowDistances;
-  float m_DistancesDecimalDigits;
+  int m_DistancesDecimalDigits;
   bool m_ShowAngles;
   bool m_ShowDistantLines;
   int  m_LineWidth;
